ASM-5133 treat empty vlan string as empty list
Previously if an empty string was passed as tagged_vlan or
untagged_vlan for force10_interface the module would try to tag or
untag "" as the vlan rather than treating it as "no vlan". It still
had the intended behavior of removing other vlans and wouldn't fail
because this module currently pretty much eats all errors. That will
have to be fixed later...