VXFM-9193 Handle long names in LLDP neighbors data
In some cases the "Rem Host Name" field value is longer than the
column width. In that case it gets ellipsized and there is no space
between it and the next column:

    Loc PortID     Rem Host Name     Rem Port Id                        Rem Chassis Id
    --------------------------------------------------------------------------------
    Te 1/3         localhost.local...e4:1d:2d:bf:d7:e0                  e4:1d:2d:bf:d7:e0

This was breaking the existing parsing which was rather ad-hoc. This
change uses a helper method in dell-asm-util to parse the data safely.